HID: core - process the Resolution Multiplier
The Resolution Multiplier is a feature report that modifies the value of
Usages within the same Logical Collection. Where set, the effective multiplier
(calculated based on the physical dimensions of the multiplier field) affects
the event value.

This was introduced for high-resolution scrolling in Windows Vista and is
commonly used on Microsoft mice.

The recommendation for the resolution multiplier is to be 1 by default.

It's unclear whether this is is used for anything but scroll wheels, so this
patch adds the parsing but doesn't actually use the resolution multiplier for
any events yet.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>