HID: core - store the collections as a basic tree
For each collection parsed, store a pointer to the parent collection (if any)
and its children into a list. This makes it a lot easier to look up which
collection(s) any given item is part of.

Notable: the top-level collection(s) are not stored in any list, to iterate
through those you will still need to run through all collections, checking for
those with parent NULL.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>