HID: input - use the Resolution Multiplier for high-resolution scrolling
Microsoft uses a magic number of 120 for a wheel click. High-resolution
scroll wheels are supposed to use a fraction of 120 to signal smaller scroll
steps. This is implemented by the Resolution Multiplier in the device.

If the multiplier is present in the report descriptor, set it to the logical
max and then use the resolution multiplier to feed high-resolution events.

To make userspace simpler, always set the REL_WHEEL_HI_RES bit. Where the
device doesn't support actual high-resolution scrolling, the value for the
high-res data will simply be a multiple of 120 every time.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>